+++ apps/easypid/assets/paradym/adaptive-icon.png
