+++ apps/easypid/assets/paradym/splash.png
