+++ apps/easypid/funke.eas.json
@@ -1,0 +1,61 @@
+{
+	"build": {
+		"base": {
+			"node": "20.11.1"
+		},
+		"development": {
+			"extends": "base",
+			"developmentClient": true,
+			"distribution": "internal",
+			"android": {
+				"buildType": "apk"
+			},
+			"ios": {
+				"simulator": true,
+				"image": "latest"
+			},
+			"env": {
+				"APP_VARIANT": "development"
+			}
+		},
+		"preview": {
+			"extends": "base",
+			"distribution": "internal",
+			"env": {
+				"APP_VARIANT": "preview"
+			}
+		},
+		"production": {
+			"extends": "base",
+			"autoIncrement": true,
+			"distribution": "store",
+			"android": {
+				"buildType": "app-bundle",
+				"resourceClass": "large"
+			}
+		}
+	},
+	"cli": {
+		"appVersionSource": "remote"
+	},
+	"submit": {
+		"production": {
+			"ios": {
+				"companyName": "Animo Solutions",
+				"ascAppId": "6636489314"
+			},
+			"android": {
+				"track": "internal"
+			}
+		},
+		"preview": {
+			"ios": {
+				"companyName": "Animo Solutions",
+				"ascAppId": "6636489314"
+			},
+			"android": {
+				"track": "internal"
+			}
+		}
+	}
+}
