+++ apps/easypid/src/app/(app)/notifications/didcomm.tsx
@@ -1,0 +1,5 @@
+import { DidCommNotificationScreen } from '@easypid/features/didcomm/DidCommNotificationScreen'
+
+export default function Screen() {
+  return <DidCommNotificationScreen />
+}
